Added support for BH1750, HTU21D and MCP9808
BH1750 : Ambient Light Sensor
HTU21D : Humidity and Temperature Sensor
MCP9808 : Temperature Sensor
Now using library lib-i2c

+++ lib-bob/src/d8x7segment.c
@@ -53,7 +53,7 @@ void d8x7segment_cls(const device_info_t *device_info) {
 
 	do {
 		max7219_spi_write_reg(device_info, i, MAX7219_CHAR_BLANK);
-	} while (--i > 0);
+	} while (--i > (uint8_t) 0);
 }
 
 /**
